+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 ### Changed
   * ENV variable to set credentials `GCP_PUBSUB_KEY_FILE_NAME`--> `GOOGLE_APPLICATION_CREDENTIALS` (default for GCP service)
+  * GqlSingleTableInheritanceFactory returns () => ClassType<Tgql> instead of just ClassType<TGql>
 ### Fixed
 ### Deprecated
 ### Security